+++ packages/theme/src/shortcuts/Popover.ts
@@ -1,0 +1,5 @@
+export const Popover = {
+  'lt-popover': 'px-1.5 py-2',
+  'lt-popover-title': 'font-600 mb-2 text-sm',
+  'lt-popover-content': 'text-sm'
+}