+++ packages/theme/src/shortcuts/Tooltip.ts
@@ -1,0 +1,4 @@
+export const Tooltip = {
+  'lt-tooltip': 'inline-block align-middle',
+  'lt-tooltip-content': 'text-xs bg-white rounded-md px-1.5 py.5 border border-solid border-gray-200 z-999'
+}